Refactoring: Create doctype-specific Python files to break up messy senstech.utils and senstech.senstech.utils
