+++ senstech/fixtures/custom_scripts/Communication.js
@@ -1,14 +1,14 @@
 frappe.ui.form.on('Communication', {
 	refresh(frm) {
         frm.add_custom_button('Unlink E-Mail Queue', function() {
-			unlinke_email_queue(frm);
+			unlink_email_queue(frm);
 	    });
     }
 })
 
-function unlinke_email_queue(frm) {
+function unlink_email_queue(frm) {
     frappe.call({
-        "method": "senstech.utils.unlinke_email_queue",
+        "method": "senstech.scripts.communication_tools.unlink_email_queue",
         "args": {
             "communication": cur_frm.doc.name
         }
