+++ senstech/fixtures/custom_scripts/Quotation.js
@@ -9,13 +9,13 @@ frappe.ui.form.on('Quotation', {
     },
     refresh(frm) {
         if (cur_frm.doc.customer_address && cur_frm.doc.shipping_address_name) {
-            update_adress_display(frm, ['address_display', 'shipping_address'], [cur_frm.doc.customer_address, cur_frm.doc.shipping_address_name], true);
+            update_address_display(frm, ['address_display', 'shipping_address'], [cur_frm.doc.customer_address, cur_frm.doc.shipping_address_name], true);
         } else {
             if (cur_frm.doc.customer_address) {
-                update_adress_display(frm, 'address_display', cur_frm.doc.customer_address);
+                update_address_display(frm, 'address_display', cur_frm.doc.customer_address);
             }
             if (cur_frm.doc.shipping_address_name) {
-                update_adress_display(frm, 'shipping_address', cur_frm.doc.shipping_address_name);
+                update_address_display(frm, 'shipping_address', cur_frm.doc.shipping_address_name);
             }
         }
     },
@@ -91,12 +91,12 @@ frappe.ui.form.on('Quotation Item', {
    }
 })
 
-function update_adress_display(frm, fields, addresses, as_list=false) {
+function update_address_display(frm, fields, addresses, as_list=false) {
     if (!as_list) {
         as_list = '';
     }
     frappe.call({
-        "method": "senstech.utils.update_adress_display",
+        "method": "senstech.scripts.tools.update_address_display",
         "args": {
             "doctype": cur_frm.doctype,
             "doc_name": cur_frm.docname,
@@ -150,7 +150,7 @@ function fetch_templates_from_party(frm) {
 
 function attach_pdf_print(frm) {
     frappe.call({
-        "method": "senstech.utils.add_freeze_pdf_to_dt",
+        "method": "senstech.scripts.tools.add_freeze_pdf_to_dt",
         "args": {
             "dt": cur_frm.doctype,
             "dn": cur_frm.docname,
@@ -164,7 +164,7 @@ function attach_pdf_print(frm) {
 
 function add_cancelled_watermark(frm) {
     frappe.call({
-        "method": "senstech.utils.add_cancelled_watermark",
+        "method": "senstech.scripts.tools.add_cancelled_watermark",
         "args": {
             "dt": cur_frm.doctype,
             "dn": cur_frm.docname
