+++ senstech/fixtures/custom_scripts/Sales Order.js
@@ -36,13 +36,13 @@ frappe.ui.form.on('Sales Order', {
     },
     refresh(frm) {
         if (cur_frm.doc.customer_address && cur_frm.doc.shipping_address_name) {
-            update_adress_display(frm, ['address_display', 'shipping_address'], [cur_frm.doc.customer_address, cur_frm.doc.shipping_address_name], true);
+            update_address_display(frm, ['address_display', 'shipping_address'], [cur_frm.doc.customer_address, cur_frm.doc.shipping_address_name], true);
         } else {
             if (cur_frm.doc.customer_address) {
-                update_adress_display(frm, 'address_display', cur_frm.doc.customer_address);
+                update_address_display(frm, 'address_display', cur_frm.doc.customer_address);
             }
             if (cur_frm.doc.shipping_address_name) {
-                update_adress_display(frm, 'shipping_address', cur_frm.doc.shipping_address_name);
+                update_address_display(frm, 'shipping_address', cur_frm.doc.shipping_address_name);
             }
         }
         setTimeout(function(){
@@ -124,7 +124,7 @@ frappe.ui.form.on('Sales Order Item', {
 
 function calculate_versanddatum(frm) {
     frappe.call({
-        "method": "senstech.utils.calculate_versanddatum",
+        "method": "senstech.scripts.sales_order_tools.calculate_versanddatum",
         "args": {
             "so": cur_frm.doc.name
         },
@@ -178,12 +178,12 @@ function reload_contacts(frm) {
 }
 
 
-function update_adress_display(frm, fields, addresses, as_list=false) {
+function update_address_display(frm, fields, addresses, as_list=false) {
     if (!as_list) {
         as_list = '';
     }
     frappe.call({
-        "method": "senstech.utils.update_adress_display",
+        "method": "senstech.scripts.tools.update_address_display",
         "args": {
             "doctype": cur_frm.doctype,
             "doc_name": cur_frm.docname,
@@ -208,7 +208,7 @@ function update_adress_display(frm, fields, addresses, as_list=false) {
 
 function attach_pdf_print(frm) {
     frappe.call({
-        "method": "senstech.utils.add_freeze_pdf_to_dt",
+        "method": "senstech.scripts.tools.add_freeze_pdf_to_dt",
         "args": {
             "dt": cur_frm.doctype,
             "dn": cur_frm.docname,
@@ -223,7 +223,7 @@ function attach_pdf_print(frm) {
 
 function add_cancelled_watermark(frm) {
     frappe.call({
-        "method": "senstech.utils.add_cancelled_watermark",
+        "method": "senstech.scripts.tools.add_cancelled_watermark",
         "args": {
             "dt": cur_frm.doctype,
             "dn": cur_frm.docname
