+++ senstech/hooks.py
@@ -36,10 +36,10 @@ app_include_js = "/assets/senstech/js/senstech.js"
 
 jenv = {
 	"methods": [
-		"get_stock_qty:senstech.utils.check_for_batch_quick_stock_entry",
-        "json_loads:senstech.utils.json_loads",
-        "template_exists:senstech.utils.template_exists",
-        "get_item_variant_description:senstech.utils.get_item_variant_description"
+		"get_stock_qty:senstech.scripts.tools.check_for_batch_quick_stock_entry",
+        "json_loads:senstech.scripts.tools.json_loads",
+        "template_exists:senstech.scripts.tools.template_exists",
+        "get_item_variant_description:senstech.scripts.item_tools.get_item_variant_description"
 	]
 }
